feat: More separation of concerns.
