+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/NativeRender.vcxproj.filters
@@ -47,8 +47,5 @@
     <ClCompile Include="pch.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="string_helpers.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
 </Project>
