+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/config.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "json.h"
 
 namespace playeveryware::eos::config
 {
@@ -144,4 +145,39 @@ namespace playeveryware::eos::config
         }
 
     };
+
+    inline static LogLevelConfig log_config_from_json_value(json_value_s* config_json)
+    {
+        struct json_object_s* config_json_object = json_value_as_object(config_json);
+        struct json_object_element_s* iter = config_json_object->start;
+        LogLevelConfig log_config;
+
+        while (iter != nullptr)
+        {
+            if (!strcmp("LogCategoryLevelPairs", iter->name->string))
+            {
+                json_array_s* pairs = json_value_as_array(iter->value);
+                for (auto e = pairs->start; e != nullptr; e = e->next)
+                {
+                    struct json_object_s* pairs_json_object = json_value_as_object(e->value);
+                    struct json_object_element_s* pairs_iter = pairs_json_object->start;
+                    while (pairs_iter != nullptr)
+                    {
+                        if (!strcmp("Category", pairs_iter->name->string))
+                        {
+                            log_config.category.push_back(json_value_as_string(pairs_iter->value)->string);
+                        }
+                        else if (!strcmp("Level", pairs_iter->name->string))
+                        {
+                            log_config.level.push_back(json_value_as_string(pairs_iter->value)->string);
+                        }
+                        pairs_iter = pairs_iter->next;
+                    }
+                }
+            }
+            iter = iter->next;
+        }
+
+        return log_config;
+    }
 }
