+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/dllmain.cpp
@@ -29,8 +29,6 @@
 #include <logging.h>
 #include <json_helpers.h>
 
-#include "json.h"
-
 using namespace playeveryware::eos::string_helpers;
 using namespace playeveryware::eos::logging;
 using namespace playeveryware::eos::config;
@@ -102,8 +100,6 @@ extern "C"
     void __declspec(dllexport) __stdcall UnityPluginUnload();
 }
 
-
-
 static const char* pick_if_32bit_else(const char* choice_if_32bit, const char* choice_if_else)
 {
 #if PLATFORM_32BITS
@@ -112,7 +108,6 @@ static const char* pick_if_32bit_else(const char* choice_if_32bit, const char* c
     return choice_if_else;
 #endif
 }
-
 
 typedef void (*log_flush_function_t)(const char* str);
 DLL_EXPORT(void) global_log_flush_with_function(log_flush_function_t log_flush_function)
@@ -142,96 +137,6 @@ EXTERN_C void EOS_CALL eos_log_callback(const EOS_LogMessage* message)
         global_logf("%s (%s): %s", message->Category, eos_loglevel_to_print_str(message->Level), message->Message);
     }
 
-}
-
-//-------------------------------------------------------------------------
-inline static uint64_t json_value_as_uint64(json_value_s* value, uint64_t default_value = 0)
-{
-    uint64_t val = 0;
-    json_number_s* n = json_value_as_number(value);
-
-    if (n != nullptr)
-    {
-        char* end = nullptr;
-        val = strtoull(n->number, &end, 10);
-    }
-    else
-    {
-        // try to treat it as a string, then parse as long
-        char* end = nullptr;
-        json_string_s* val_as_str = json_value_as_string(value);
-        if (val_as_str == nullptr || strlen(val_as_str->string) == 0)
-        {
-            val = default_value;
-        }
-        else
-        {
-            val = strtoull(val_as_str->string, &end, 10);
-        }
-    }
-
-    return val;
-}
-
-
-//-------------------------------------------------------------------------
-inline static uint32_t json_value_as_uint32(json_value_s* value, uint32_t default_value = 0)
-{
-    uint32_t val = 0;
-    json_number_s* n = json_value_as_number(value);
-
-    if (n != nullptr)
-    {
-        char* end = nullptr;
-        val = strtoul(n->number, &end, 10);
-    }
-    else
-    {
-        // try to treat it as a string, then parse as long
-        char* end = nullptr;
-        json_string_s* val_as_str = json_value_as_string(value);
-
-        if (val_as_str == nullptr || strlen(val_as_str->string) == 0)
-        {
-            val = default_value;
-        }
-        else
-        {
-            val = strtoul(val_as_str->string, &end, 10);
-        }
-    }
-
-    return val;
-}
-
-//-------------------------------------------------------------------------
-inline static double json_value_as_double(json_value_s* value, double default_value = 0.0)
-{
-    double val = 0.0;
-    json_number_s* n = json_value_as_number(value);
-
-    if (n != nullptr)
-    {
-        char* end = nullptr;
-        val = strtod(n->number, &end);
-    }
-    else
-    {
-        // try to treat it as a string, then parse as long
-        char* end = nullptr;
-        json_string_s* val_as_str = json_value_as_string(value);
-
-        if (val_as_str == nullptr || strlen(val_as_str->string) == 0)
-        {
-            val = default_value;
-        }
-        else
-        {
-            val = strtod(val_as_str->string, &end);
-        }
-    }
-
-    return val;
 }
 
 //-------------------------------------------------------------------------
@@ -384,28 +289,6 @@ static char* GetCacheDirectory()
     }
 
     return s_tempPathBuffer;
-}
-
-//-------------------------------------------------------------------------
-static json_value_s* read_config_json_as_json_from_path(std::filesystem::path path_to_config_json)
-{
-    log_inform(("json path" + to_utf8_str(path_to_config_json)).c_str());
-    uintmax_t config_file_size = std::filesystem::file_size(path_to_config_json);
-    if (config_file_size > SIZE_MAX)
-    {
-        throw std::filesystem::filesystem_error("File is too large", std::make_error_code(std::errc::file_too_large));
-    }
-
-    FILE* file = nullptr;
-    errno_t config_file_error = _wfopen_s(&file, path_to_config_json.wstring().c_str(), L"r");
-    char* buffer = (char*)calloc(1, static_cast<size_t>(config_file_size));
-
-    size_t bytes_read = fread(buffer, 1, static_cast<size_t>(config_file_size), file);
-    fclose(file);
-    struct json_value_s* config_json = json_parse(buffer, bytes_read);
-    free(buffer);
-
-    return config_json;
 }
 
 //-------------------------------------------------------------------------
@@ -572,40 +455,7 @@ static EOSConfig eos_config_from_json_value(json_value_s* config_json)
     return eos_config;
 }
 
-static LogLevelConfig log_config_from_json_value(json_value_s* config_json)
-{
-    struct json_object_s* config_json_object = json_value_as_object(config_json);
-    struct json_object_element_s* iter = config_json_object->start;
-    LogLevelConfig log_config;
 
-    while (iter != nullptr)
-    {
-        if (!strcmp("LogCategoryLevelPairs", iter->name->string))
-        {
-            json_array_s* pairs = json_value_as_array(iter->value);
-            for (auto e = pairs->start; e != nullptr; e = e->next)
-            {
-                struct json_object_s* pairs_json_object = json_value_as_object(e->value);
-                struct json_object_element_s* pairs_iter = pairs_json_object->start;
-                while (pairs_iter != nullptr)
-                {
-                    if (!strcmp("Category", pairs_iter->name->string))
-                    {
-                        log_config.category.push_back(json_value_as_string(pairs_iter->value)->string);
-                    }
-                    else if (!strcmp("Level", pairs_iter->name->string))
-                    {
-                        log_config.level.push_back(json_value_as_string(pairs_iter->value)->string);
-                    }
-                    pairs_iter = pairs_iter->next;
-                }
-            }
-        }
-        iter = iter->next;
-    }
-    
-    return log_config;
-}
 
 template<typename T>
 static T collect_flags(const std::map<std::string, T>* strings_to_enum_values, T default_value, json_object_element_s* iter)
